Split rpc into subpackage
